some docs and more refactoring
