+++ src/cache/interop.ts
@@ -1,7 +1,11 @@
 import { EventApi, EventInput } from "@fullcalendar/core";
-import { OFCEvent } from "./types";
+import { OFCEvent } from "../types";
 
 import { DateTime, Duration } from "luxon";
+
+/*
+ * Functions for converting between the types used by the FullCalendar view plugin and types used internally by Obsidian Full Calendar.
+ */
 
 const parseTime = (time: string): Duration | null => {
 	let parsed = DateTime.fromFormat(time, "h:mm a");
